PET-103 refactor : User-Application Dto 패키지 구조 통일
