+++ Domain-Module/User-Module/User-Application/src/main/java/com/pawith/usermodule/dto/request/UserNicknameChangeRequest.java
@@ -1,4 +1,4 @@
-package com.pawith.usermodule.service.dto;
+package com.pawith.usermodule.dto.request;
 
 import lombok.AccessLevel;
 import lombok.Getter;
