+++ Domain-Module/User-Module/User-Application/src/main/java/com/pawith/usermodule/dto/response/UserAuthorityInfoResponse.java
@@ -1,4 +1,4 @@
-package com.pawith.usermodule.service.dto;
+package com.pawith.usermodule.dto.response;
 
 import com.pawith.usermodule.entity.Authority;
 import lombok.Getter;
