+++ Domain-Module/User-Module/User-Application/src/main/java/com/pawith/usermodule/service/UserAuthorityGetUseCase.java
@@ -1,7 +1,7 @@
 package com.pawith.usermodule.service;
 
 import com.pawith.commonmodule.annotation.ApplicationService;
-import com.pawith.usermodule.service.dto.UserAuthorityInfoResponse;
+import com.pawith.usermodule.dto.response.UserAuthorityInfoResponse;
 import com.pawith.usermodule.entity.UserAuthority;
 import com.pawith.usermodule.utils.UserUtils;
 import lombok.RequiredArgsConstructor;
