+++ Domain-Module/User-Module/User-Application/src/main/java/com/pawith/usermodule/service/UserNicknameChangeUseCase.java
@@ -1,10 +1,9 @@
 package com.pawith.usermodule.service;
 
 import com.pawith.commonmodule.annotation.ApplicationService;
-import com.pawith.usermodule.service.dto.UserNicknameChangeRequest;
-import com.pawith.usermodule.entity.Authority;
 import com.pawith.usermodule.entity.User;
 import com.pawith.usermodule.entity.UserAuthority;
+import com.pawith.usermodule.dto.request.UserNicknameChangeRequest;
 import com.pawith.usermodule.utils.UserUtils;
 import lombok.RequiredArgsConstructor;
 import org.springframework.transaction.annotation.Transactional;
@@ -20,8 +19,6 @@ public class UserNicknameChangeUseCase {
         final User user = UserUtils.getAccessUser();
         user.updateNickname(request.getNickname());
         final UserAuthority userAuthority = userAuthorityQueryService.findByEmail(user.getEmail());
-        if(userAuthority.getAuthority().equals(Authority.GUEST)){
-            userAuthority.changeUserAuthority();
-        }
+        userAuthority.changeUserAuthority();
     }
 }
